refactor(test): move headless PNG helpers to utils
