+++ lib/test/headless/polar-smoke.test.ts
@@ -1,9 +1,9 @@
 import path from "node:path";
 import { fileURLToPath } from "node:url";
 import { describe, expect, it } from "vitest";
-import { compareGoldenPng } from "./compareGoldenPng";
+import { compareGoldenPng } from "./utils/compareGoldenPng";
 import { smokePolarChart, smokePolarSize } from "./fixtures/smokePolarChart";
-import { renderHeadlessToPng } from "./renderHeadlessToPng";
+import { renderHeadlessToPng } from "./utils/renderHeadlessToPng";
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 const goldenPath = path.join(__dirname, "__golden__", "polar-smoke.png");
