+++ lib/test/headless/utils/compareGoldenPng.ts
