+++ lib/test/headless/utils/renderHeadlessToPng.ts
